--dev <baseUrl>
'--dev' now optionally receives an argument, which allows you to point ecp-cli
to servers other than prod and dev.

Side-effect: now the verb has to come *before* the option. Where before you could do

    --dev login

now it has to be

    login --dev

Probably something to do with that weird library (argparse).
Change it to docopt to see if you can bring the old behaviour back.